Enable *warn-on-reflection* and add missing hints
At the moment there are several reflection calls occuring on the
hot-path for request handling. This is both slow and makes it harder to
compile with GraalVM.

This commit enables the `*warn-on-reflection*` global on all namespaces
and adds the missing type hints to locations where reflection were
occuring.

In some cases this required restructuring code a bit to better apply the
hints.